Refactor Names and add README
